+++ ntest_timeout/src/lib.rs
@@ -61,7 +61,7 @@ pub fn timeout(attr: TokenStream, item: TokenStream) -> TokenStream {
             fn ntest_callback() #output
             #body
             let ntest_timeout_now = std::time::Instant::now();
-            match ntest::execute_with_timeout(&ntest_callback, #time_ms as u64) {
+            match ntest_proc_macro_helper::execute_with_timeout(&ntest_callback, #time_ms as u64) {
                 Some(result) => return result,
                 None => panic!("timeout: the function call took {} ms. Max time {} ms", ntest_timeout_now.elapsed().as_millis(), #time_ms),
             }